requirements.txt
# pip install -r requirements.txt
boto3==1.13.7
botocore==1.16.7
Flask==1.1.2
psycopg2-binary==2.8.6